requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-WTF==1.1.1
python-dotenv==1.0.1
google-genai==1.46.0





